requirements.txt
# Core Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# Streamlit
streamlit==1.40.0

# LangChain & LangGraph (Updated versions)
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.3.0
langgraph==0.2.45
langchain-community==0.3.7
langchain-core==0.3.17
langchain-text-splitters==0.3.11

# MCP SDK (Note: We're using custom MCP servers, not the official SDK)
# The mcp_servers/ directory contains our custom implementation

# Code Analysis
gitpython==3.1.43
pygments==2.18.0
radon==6.0.1

# GitHub API
PyGithub==2.5.0
requests==2.32.3

# File Processing
python-multipart==0.0.12
aiofiles==24.1.0

# Utilities
httpx==0.27.2
python-json-logger==3.2.1
